Add `Host::borrow_game` and `Host::borrow_game_mut`
This change introduces another way of accessing the game state for a `Host` via `borrow_game` and `borrow_game_mut`. This may be more straightforward than `with_game` and `with_game_mut` when accessing state for very short periods of time. The caveat is that using `with_game` and `with_game_mut` can make the lifetime of the transaction less clear to the reader by introducing `Ref`/`RefMut` guards.