Allow `Play::handle_event` to mutate `Event` argument
The intention is to allow handling the event internally to modify the `Event` in some way before it is yielded to the main application. For example, if a game has `type Event = Option<SomeEvent>`, the `handle_event` implementation might call `event.take()` to consume the event, preventing it from being processed externally.